status: fix branch shown when not only bisecting
In 83c750acde (wt-status.*: better advice for git status added,
2012-06-05), git-status received new informative messages to describe
the ongoing work in a worktree.

Multiple operations can be performed concurrently in a worktree.  For
example, during a rebase, the user can also perform a cherry-pick.  In
that situation, git-status only shows information about one of them,
prioritizing which one, in order: merge, am, rebase, cherry-pick.

However, when a bisect is also in progress, git-status includes, in
addition to the description of any other ongoing operation, the
description of the bisect.  This means that, in this case, it shows
description for two ongoing operations.

In 0722c805d6 (status: show the branch name if possible in in-progress
info, 2013-02-03), the messages introduced in 83c750acde were improved
to show, if possible, the branch in which the described operation was
initiated.

Unfortunately, git-status only records one branch related to one
operation.  However, in the situation described above, when a bisect is
combined with another operation, it is necessary to record two branches,
which, it is important to note, may be different.

Therefore, when that happens, we show incorrect information:

   $ git checkout -b bisect-branch
   $ git bisect start
   $ git checkout -b rebase-branch
   $ GIT_SEQUENCE_EDITOR='echo break >>' git rebase -i HEAD~
   $ git status
   interactive rebase in progress ...
   ...

   You are currently editing a commit while rebasing branch 'bisect-branch' on '...'.

   You are currently bisecting, started from branch 'bisect-branch'.

   ...

Days after 0722c805d6, this bisect circumstance hit us again; a leak was
introduced in 8b87cfd000 (wt-status: move strbuf into
read_and_strip_branch(), 2013-03-16).

Let's fix git-status to display accurate information and also fix the
leak, considering the branch where the bisect was started separately
from the branch related to other ongoing operations.

Signed-off-by: Rubén Justo <rjusto@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>