update-index: add --show-index-version
"git update-index --version N" is used to set the index format
version to a specific version, but there was no way to query the
current version used in the on-disk index file.

Teach the command a new "--show-index-version" option, and also
teach the "--index-version N" option to report what the version was
when run with the "--verbose" option.

Signed-off-by: Junio C Hamano <gitster@pobox.com>